Use #respond_to? instead of === to determine URI-ness
Persistent#request_setup has a check on the class of the value
req_or_uri which tests to see if req_or_uri is an instance of URI.
But there's a good chance that someone might pass an instance of
Addressable::URI instead of URI, which will fail this check.

Addressable::URI is basically a drop-in replacement for URI, so there's
no reason to care which of these objects are passed, just that it
responds to the #request_uri method. So, #respond_to? to the rescue!