Add dev/test dependency and fix broken test
`TestNetHttpPersistent#test_pipeline` was getting skipped by default, since `net-http-pipeline` wasn't installed as a test dependency.  However, when installed, the skipped test was failing with `NoMethodError: undefined method `start' for #<Net::HTTP::Persistent::Connection:0x007fcf0d992d28>`.

This adds `net-http-pipeline` as a dev dependency so the test isn't skipped in CI and fixes the test so it passes.