Expose `ConnectionPool#reload`
Ref: https://github.com/mperham/connection_pool/pull/140

When forking a process, you need to close existing connection
to avoid sharing them across processes. `shutdown` does that, but
it also mark the pool as no longer usable.

`connection_pool` `2.2.4` introduced `#reload` that discard existing
connections, but let the pool be used again later. It's a much better
fit for an `after_fork` callback.